Replaced MPI_DOUBLE_PRECISION with MPI_DOUBLE
I got the following error message when I ran the code on a new
machine:

An error occurred in MPI_Allreduce: the reduction operation
MPI_MIN is not defined on the MPI_DOUBLE_PRECISION datatype

Ryan explained to me that MPI_DOUBLE_PRECISION is a Fortran type,
and it was only working for us by coincidence.  He said to replace
all instances with MPI_DOUBLE, the corresponding C type.  Thanks
Ryan!