Fixed build warnings
Sidenote: -Wconversion is an incredibly useful compiler flag for
gcc.  It warns you every time you implicitly convert types if
there's a chance of overflow/loss of precision.

This should fix the bug I saw on skybridge, where a size_t was
being implicitly cast to an int, but the int wasn't big enough.