Code now frees MPI communicators
There is an (implementation-dependent) upper limit on the number
of MPI communicators that can be active at one time.  It tends to
be over 4000.  When creating tensors in a loop (for partial
reconstruction), we were hitting that limit because each tensor
created several MPI communicators and did not free them in its
destructor.