Added sequential binary IO routines and a test
Note that these routines will depend on the endian-ness of the
machine being used.  It's a bad idea to write a file on a big-
endian platform and try to read it on a little-endian platform.

I can add a flag to the IO routines to swap the endian-ness if
necessary, but for now the solution is "don't do that".