Modified Woody's stringParseSizeArray function
Our SizeArray class does not allow for empty arrays (ones with
no dimensions) and throws an exception when you try to create one.
If stringParseSizeArray fails to find the specified string in the
input file, it tries to return an empty array...only it can't, so
an exception is thrown.  This means that if the user forgets to
specify a required parameter, they get a confusing error message
about how SizeArray cannot be size 0.

Woody's function now returns NULL if the parameter is not found.
We can then check for null pointers in our drivers, and output a
better error message if one is found.  (Global dimensions is null?
I guess the user forgot to specify it.  Print "Error: Global
dimensions is a required parameter and must be specified.")