Added sequential full-workflow test
This test generates a random tensor using Tucker_generate.cpp,
compresses it with Tucker_sthosvd.cpp, and reconstructs it with
Tucker_reconstruct.cpp.  If the compressed tensor has incorrect
dimensions, or the reconstructed data is not approximately equal
to the original data, the test fails.

Since the test was written in python, we now have a python
dependency, which shouldn't be a big deal.  Worst case scenerio,
users with a very old version of python will have a failing test.