New Gram algorithm now requires less memory
Theoretically speaking, the new Gram algorithm should require
approximately 3x as much memory as the tensor itself.  This is
because it needs to store

* the original tensor
* the packed data for sending (if mode > 0)
* the received data
* the unpacked data (if mode < N-1)

After the data is received, we can deallocate the memory for the
send buffer.  We were not previously doing this, so the code was
requiring 4x as much memory as the tensor.