Only use the single reduce-scatter on the host
The pack phase is very expensive on the device.  I think this issue is
it is calling many small kernels without enough parallelism.  A single
kernel across all of these little pack kernels would be better, but is
nontrivial to write.  So we'll just use the multiple reduction approach
(which doesn't pack) for now (which also saves limited GPU memory).